LayoutTests/platform/mac/accessibility/search-when-element-starts-in-table-expected.txt

Column
heading

outside heading

This tests that when we start searching from an element within a table, we do not get as a result, the same element.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS resultElement.role is 'AXRole: AXHeading'
PASS resultElement.title is 'AXTitle: outside heading'
PASS resultElement.role is 'AXRole: '
PASS resultElement.role is 'AXRole: AXCell'
PASS successfullyParsed is true

TEST COMPLETE
